+++ src/test/java/com/gotreaux/aoc/puzzles/year2019/day1/RocketEquationPuzzleTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 class RocketEquationPuzzleTest {
     @ParameterizedTest
     @MethodSource("provideFuelRequirement")
-    void fuelRequirement(String input, long expectedFuelRequirement) throws Exception {
+    void fuelRequirement(String input, int expectedFuelRequirement) throws Exception {
         StringInputProvider inputProvider = new StringInputProvider(input);
 
         RocketEquationPuzzle puzzle = new RocketEquationPuzzle(inputProvider);
@@ -21,7 +21,7 @@ class RocketEquationPuzzleTest {
 
     @ParameterizedTest
     @MethodSource("provideAdditionalFuelRequirement")
-    void additionalFuelRequirement(String input, long expectedFuelRequirement) throws Exception {
+    void additionalFuelRequirement(String input, int expectedFuelRequirement) throws Exception {
         StringInputProvider inputProvider = new StringInputProvider(input);
 
         RocketEquationPuzzle puzzle = new RocketEquationPuzzle(inputProvider);
@@ -31,17 +31,17 @@ class RocketEquationPuzzleTest {
 
     private static Stream<Arguments> provideFuelRequirement() {
         return Stream.of(
-                Arguments.of("12", 2L),
-                Arguments.of("14", 2L),
-                Arguments.of("1969", 654L),
-                Arguments.of("100756", 33583L));
+                Arguments.of("12", 2),
+                Arguments.of("14", 2),
+                Arguments.of("1969", 654),
+                Arguments.of("100756", 33583));
     }
 
     private static Stream<Arguments> provideAdditionalFuelRequirement() {
         return Stream.of(
-                Arguments.of("12", 2L),
-                Arguments.of("14", 2L),
-                Arguments.of("1969", 966L),
-                Arguments.of("100756", 50346L));
+                Arguments.of("12", 2),
+                Arguments.of("14", 2),
+                Arguments.of("1969", 966),
+                Arguments.of("100756", 50346));
     }
 }
